Fix missing casts for `size()`/`int` comparison

+++ src/view_control.cpp
@@ -1686,8 +1686,8 @@ void View::DrawDynamSymbolOnly(DeviceContext *dc, Staff *staff, Dynam *dynam, co
     // combination of different dynamics glyphs we need to get total width of all elements
     const int left = m_doc->GetGlyphLeft(dynamSymbol.at(0), staff->m_drawingStaffSize, false);
     int width = 0;
-    for (int i = 0; i < dynamSymbol.size(); ++i) {
-        if (i == dynamSymbol.size() - 1) {
+    for (int i = 0; i < (int)dynamSymbol.size(); ++i) {
+        if (i == (int)dynamSymbol.size() - 1) {
             width += m_doc->GetGlyphRight(dynamSymbol.at(i), staff->m_drawingStaffSize, false);
         }
         else {
